gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
python-jose